+++ DKBasicEngine 1.0/Core/Components/Physics.cs
@@ -31,7 +31,7 @@ namespace DKBasicEngine_1_0
 
         public void Jump()
         {
-            int StartPositon = Parent.Y;
+            double StartPositon = Parent.Y;
             int MinJump = 10;
 
             do
