+++ DKBasicEngine 1.0/Core/Interfaces/I3Dimensional.cs
@@ -8,13 +8,13 @@ namespace DKBasicEngine_1_0
 {
     public interface I3Dimensional
     {
-        int width { get; }
-        int height { get; }
-        int depth { get; }
+        double width { get; }
+        double height { get; }
+        double depth { get; }
 
-        int X { get; set; }
-        int Y { get; set; }
-        int Z { get; set; }
+        double X { get; set; }
+        double Y { get; set; }
+        double Z { get; set; }
 
         double ScaleX { get; set; }
         double ScaleY { get; set; }
